+++ all-is-cubes/src/block/modifier/zoom.rs
@@ -51,7 +51,7 @@ impl Zoom {
 
     /// Decompose into parts, for serialization.
     #[cfg(feature = "save")]
-    pub(crate) fn to_serial_schema(&self) -> crate::save::schema::ModifierSer {
+    pub(crate) fn to_serial_schema(&self) -> crate::save::schema::ModifierSer<'static> {
         let Zoom { scale, offset } = *self;
         crate::save::schema::ModifierSer::ZoomV1 {
             scale,